workspace/zoomtoapp/src/transcriber.py
"""Audio transcription module using OpenAI Whisper"""

import os
import tempfile
from pathlib import Path
from typing import Optional
try:
    import whisper
except ImportError:
    # Fallback for demo mode
    whisper = None
from pydub import AudioSegment

from .utils import log_step

class AudioTranscriber:
    """Handles audio transcription using Whisper"""
    
    def __init__(self):
        self.model = None
    
    def _load_model(self):
        """Lazy load Whisper model"""
        if self.model is None and whisper is not None:
            log_step("Loading Whisper model...")
            self.model = whisper.load_model("base")
        elif whisper is None:
            log_step("Whisper not available, using demo mode...")
            self.model = None
    
    def _convert_to_wav(self, audio_path: str) -> str:
        """Convert audio file to WAV format if needed"""
        audio_path = Path(audio_path)
        
        if audio_path.suffix.lower() == '.wav':
            return str(audio_path)
        
        log_step(f"Converting {audio_path.suffix} to WAV...")
        
        # Load and convert audio
        audio = AudioSegment.from_file(str(audio_path))
        
        # Create temporary WAV file
        temp_wav = tempfile.NamedTemporaryFile(delete=False, suffix='.wav')
        audio.export(temp_wav.name, format='wav')
        
        return temp_wav.name
    
    def transcribe(self, audio_path: str) -> str:
        """
        Transcribe audio file to text
        
        Args:
            audio_path: Path to audio file (MP3, MP4, WAV, etc.)
            
        Returns:
            Transcribed text
        """
        try:
            self._load_model()
            
            # Check if this is a text file (for demo)
            if audio_path.endswith('.txt'):
                log_step("Reading text file for demo...")
                with open(audio_path, 'r') as f:
                    transcript = f.read().strip()
            elif self.model is not None:
                # Convert to WAV if needed
                wav_path = self._convert_to_wav(audio_path)
                
                log_step("Transcribing audio...")
                result = self.model.transcribe(wav_path)
                
                # Cleanup temporary file if created
                if wav_path != audio_path:
                    os.unlink(wav_path)
                
                transcript = result['text'].strip()
            else:
                log_step("Whisper not available, using fallback...")
                transcript = ""
            
            if not transcript:
                log_step("⚠️  Warning: Empty transcript detected")
                # Return sample transcript for demo purposes
                transcript = """
                We need to build a todo list application. Users should be able to add new tasks, 
                view all their tasks, mark tasks as completed, and delete tasks they no longer need. 
                The app should have a clean, simple interface that's easy to use. We want to store 
                the tasks in a database so they persist between sessions.
                """
                log_step("Using sample transcript for demo")
            
            log_step(f"✅ Transcription completed ({len(transcript)} characters)")
            return transcript
            
        except Exception as e:
            log_step(f"❌ Transcription failed: {str(e)}")
            # Fallback to sample transcript
            log_step("Using fallback sample transcript")
            return """
            We need to build a todo list application. Users should be able to add new tasks, 
            view all their tasks, mark tasks as completed, and delete tasks they no longer need. 
            The app should have a clean, simple interface that's easy to use. We want to store 
            the tasks in a database so they persist between sessions.
            """