workspace/zoomtoapp/requirements.txt
openai>=1.3.0
langchain>=0.0.350
langchain-openai>=0.0.2
openai-whisper>=20231117
torch>=2.1.0
torchaudio>=2.1.0
requests>=2.31.0
GitPython>=3.1.40
python-dotenv>=1.0.0
click>=8.1.7
pydub>=0.25.1
ffmpeg-python>=0.2.0
pytest>=7.4.3
coverage>=7.3.2